Api set profile image

+++ Nexus.OAuth.Api/Controllers/FilesController.cs
@@ -67,5 +67,5 @@ public class FilesController : ApiController
 
     [NonAction]
     private static DirectoryType DirectoryByResourceType(ResourceType resourceType)
-        => (DirectoryType)((int)resourceType);
+        => (DirectoryType)(resourceType);
 }
